Initial commit to do microversions
With the changes in this commit, you can specify the version of
client you wish to use (1.0 or 1.1)  The client will get the
best match for the server/client and instantiate that time of client.

If no version is specified, the client will use the max it supports
(assuming server supports it).

In this commit, we'll set the client to ask for the 1.0 version
so that we do not break on the secrets functional tests.  In the
subsequent patch where we add the secrets consumer column support,
we'll change to use the MAX_VERSION.

Change-Id: I2a0a5315daca8ce5bd6164cebbd2f917348c7675